chore: refactor basic Python and Pandas data types handling
- provide CoreObjectsBackend to handle basic Python and Pandas data types
- remove built-in mllib support, keeping SparkBackend for reference
- simplify base.OmegaStore implementation, delegating all storage to backends (other than Metadata)
- refactor MDREGISTY.KINDS to be empty by default, accepting only dynamically registered backends
- refactor OmegaJobs as a backend
