+++ omegaml/tests/util.py
@@ -1,9 +1,8 @@
-from http import HTTPStatus
-
 import logging
 import os
 import sys
 import warnings
+from http import HTTPStatus
 
 from omegaml import Omega
 from omegaml.client.lunamon import LunaMonitor
@@ -42,7 +41,7 @@ class OmegaTestMixin(object):
                   **drop_kwargs) for m in part.list(hidden=True, include_temp=True, raw=True)]
             # ignore system members, as they may get recreated e.g. by LunaMonitor
             existing = [m.name for m in part.list(hidden=True, include_temp=True, raw=True)
-                        if not '.system' in m.name.startswith()]
+                        if not '.system' in m.name]
             self.assertListEqual(existing, [])
 
     @property
